Implement new Header component
